create user room package and move related classes to them
