+++ core/src/main/java/hu/uni/miskolc/iit/sweng/hotelReservation/model/Reservation.java
@@ -1,11 +1,13 @@
 package hu.uni.miskolc.iit.sweng.hotelReservation.model;
 
+import hu.uni.miskolc.iit.sweng.hotelReservation.model.room.Room;
+
 import java.util.Collection;
 
 public class Reservation {
 
     private Collection<Room> Rooms;
-    private User User;
+    private hu.uni.miskolc.iit.sweng.hotelReservation.model.user.User User;
