+++ core/src/main/java/hu/uni/miskolc/iit/sweng/hotelReservation/model/user/Nationality.java
@@ -1,4 +1,4 @@
-package hu.uni.miskolc.iit.sweng.hotelReservation.model;
+package hu.uni.miskolc.iit.sweng.hotelReservation.model.user;
 
 /**
  * created By Kaikoni Basil 19-10-2017
