+++ core/src/main/java/hu/uni/miskolc/iit/sweng/hotelReservation/service/UserManagerService.java
@@ -1,6 +1,6 @@
 package hu.uni.miskolc.iit.sweng.hotelReservation.service;
-import hu.uni.miskolc.iit.sweng.hotelReservation.model.Nationality;
-import hu.uni.miskolc.iit.sweng.hotelReservation.model.User;
+import hu.uni.miskolc.iit.sweng.hotelReservation.model.user.Nationality;
+import hu.uni.miskolc.iit.sweng.hotelReservation.model.user.User;
 import hu.uni.miskolc.iit.sweng.hotelReservation.service.exceptions.UserAlreadyExistException;
 import hu.uni.miskolc.iit.sweng.hotelReservation.service.exceptions.UserNotFoundException;
